Update README.md
A virtual model of a genuine system can be created with various benefits, particularly during the design stages, before the system is actually manufactured. When testing or deciding what has to be improved in the prototype, simulating a system might be useful. modifications caused by the existence of flaws. Here, taking advantage of a game's characteristics primarily used in video games as an engine to simulate actual prototypes provides fresh insights into the research. sophisticated rendering powers, physical simulation, this setting very accurate in describing the many materials that are available. Appealing outside the realm of video games. An additional crucial feature that is obtainable in these simulators is the amalgamation of precisely rendered physical characteristics and the visual representation of physical actions. When combined, they produce a visually realistic and physically accurate representation. This article, which is mostly based on the C++ language and BluePrint, explores the creation of behavioral models inside the Unreal Engine gaming environment. The developed models will be applied to study on improvement, performance monitoring, and design analysis. A digital twin of a DC motor that has been simulated under both ideal and imperfect operating circumstances has been selected as the case study model.